pyproject.toml
[tool.poetry]
name = "ayon_server"
version = "1.0.0"
description = ""
authors = ["Ynput.io <info@ynput.io>"]

[tool.poetry.dependencies]
python = "^3.10"
nxtools = "^1.6"
orjson = "^3.6.7"
fastapi = "^0.73.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.97.0"}
uvicorn = {extras = ["standard"], version = "^0.17.5"}
aioredis = "^2.0.1"
asyncpg = "^0.25.0"
email-validator = "^1.1.3"
typer = "^0.4.1"
yaoauth2 = "^1.0.0"
toml = "^0.10.2"
types-toml = "^0.10.8"
psutil = "^5.9.4"
pre-commit = "^2.20.0"
codenamize = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-order = "^1.0.1"
black = {version = "^22.1.0", allow-prereleases = true}
pre-commit = "^2.17.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.991"

[tool.isort]
profile = "black"
skip = ["addons/"]

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

